Reworking

+++ backend/controllers/cards.js
@@ -1,43 +1,53 @@
+const BadRequestError = require('../middlewares/errors/BadRequestError');
+const ForbiddenError = require('../middlewares/errors/ForbiddenError');
+const NotFoundError = require('../middlewares/errors/NotFoundError');
 const Card = require('../models/card');
 
-const getCards = async (req, res) => {
+const getCards = async (req, res, next) => {
   try {
     const cards = await Card.find({});
     return res.status(200).send(cards);
   } catch (err) {
-    return res.status(500).send({ message: 'Ошибка на сервере' });
+    return next();
   }
 };
 
-const createCard = async (req, res) => {
+const createCard = async (req, res, next) => {
   try {
     const { name, link } = req.body;
     const card = await Card.create({ name, link, owner: req.user._id });
     return res.status(200).send(card);
   } catch (err) {
     if (err.name === 'ValidationError') {
-      return res.status(400).send({ message: 'Неверная ссылка' });
+      return next(new BadRequestError('Неверная ссылка'));
+      // return res.status(400).send({ message: 'Неверная ссылка' });
     }
-    return res.status(500).send({ message: 'Ошибка на сервере' });
+    return next();
+    // return res.status(500).send({ message: 'Ошибка на сервере' });
   }
 };
 
-const deleteCard = async (req, res) => {
+const deleteCard = async (req, res, next) => {
   try {
+    const card = await Card.findById(req.params.cardId);
+    if (card.owner !== req.user._id) {
+      return next(new ForbiddenError('Недостаточно прав'));
+    }
     await Card.findByIdAndDelete(req.params.cardId).orFail();
     return res.status(200).send({ message: 'Карточка удалена' });
   } catch (err) {
     if (err.name === 'CastError') {
-      return res.status(400).send({ message: 'Переданы некорректные данные' });
+      return next(new BadRequestError('Переданы некорректные данные'));
     }
     if (err.name === 'DocumentNotFoundError') {
-      return res.status(404).send({ message: 'Нет карточки с таким id' });
+      return next(new NotFoundError('Нет карточки с таким id'));
+      // return res.status(404).send({ message: 'Нет карточки с таким id' });
     }
-    return res.status(500).send({ message: 'Ошибка на сервере' });
+    return next();
   }
 };
 
-const likeCard = async (req, res) => {
+const likeCard = async (req, res, next) => {
   try {
     const likes = await Card.findByIdAndUpdate(
       req.params.cardId,
@@ -47,16 +57,16 @@ const likeCard = async (req, res) => {
     return res.status(200).send({ data: likes });
   } catch (err) {
     if (err.name === 'CastError') {
-      return res.status(400).send({ message: 'Переданы некорректные данные' });
+      return next(new BadRequestError('Переданы некорректные данные'));
     }
     if (err.name === 'DocumentNotFoundError') {
-      return res.status(404).send({ message: 'Нет карточки с таким id' });
+      return next(new NotFoundError('Нет карточки с таким id'));
     }
-    return res.status(500).send({ message: 'Ошибка на сервере' });
+    return next();
   }
 };
 
-const dislikeCard = async (req, res) => {
+const dislikeCard = async (req, res, next) => {
   try {
     const likes = await Card.findByIdAndUpdate(
       req.params.cardId,
@@ -66,12 +76,13 @@ const dislikeCard = async (req, res) => {
     return res.status(200).send({ data: likes });
   } catch (err) {
     if (err.name === 'CastError') {
-      return res.status(400).send({ message: 'Переданы некорректные данные' });
+      return next(new BadRequestError('Переданы некорректные данные'));
     }
     if (err.name === 'DocumentNotFoundError') {
-      return res.status(404).send({ message: 'Нет карточки с таким id' });
+      return next(new NotFoundError('Нет карточки с таким id'));
+      // return res.status(404).send({ message: 'Нет карточки с таким id' });
     }
-    return res.status(500).send({ message: 'Ошибка на сервере' });
+    return next();
   }
 };
